# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: cmp是一个客户关系管理平台
  version: 1.0.0
  title: 客户关系管理平台
basePath: /cam
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/user/list:
    get:
      tags:
        - 管理员管理
      summary: 管理员列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: username
          description: 用户名
          in: query
          type: string
      responses:
        '200':
          description: 返回管理员列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
...
  /sys/user/info/{userId}:
    get:
      tags:
        - 管理员管理
      summary: 获取管理员信息
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回管理员信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysUserEntity'
  /sys/user/password:
    post:
      tags:
        - 管理员管理
      summary: 修改密码
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/PasswordForm'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/save:
    post:
      tags:
        - 管理员管理
      summary: 添加管理员
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/update:
    post:
      tags:
        - 管理员管理
      summary: 修改管理员
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/delete:
    post:
      tags:
        - 管理员管理
      summary: 删除管理员
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/bindingExtension:
    post:
      tags:
        - 管理员管理
      summary: 绑定分机号
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /tel/crmchanneltelephoneagent/list:
    get:
      tags:
        - 渠道管理电话工号
      summary: 渠道管理电话工号列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: channelId
          description: 渠道管理id
          in: query
          type: string
        - name: telephoneAgent
          description: 话务工号
          in: query
          type: string
        - name: skillsGroup
          description: 技能组
          in: query
          type: string
        - name: processingTime
          description: 后续处理时长
          in: query
          type: integer
        - name: ringingCount
          description: 振铃次数
          in: query
          type: integer
        - name: agentNumber
          description: 绑定坐席工号
          in: query
          type: string
        - name: agentName
          description: 绑定坐席名称
          in: query
          type: string
        - name: isAnswer
          description: 是否自动接听 0:否 1:是
          in: query
          type: string
        - name: isHangupSatisfaction
          description: 是否自动挂机自动转满意度 0:否 1:是
          in: query
          type: string
        - name: isBroadcasterAgent
          description: 是否自动播报工号 0:否 1:是
          in: query
          type: string
        - name: createCode
          description: 创建人代码
          in: query
          type: string
        - name: createName
          description: 创建人
          in: query
          type: string
        - name: startCreateTime
          description: 开始创建时间
          in: query
          type: string
          format: date
        - name: endCreateTime
          description: 结束创建时间
          in: query
          type: string
          format: date
        - name: updateCode
          description: 修改人代码
          in: query
          type: string
        - name: updateName
          description: 修改人
          in: query
          type: string
        - name: startUpdateTime
          description: 开始修改时间
          in: query
          type: string
          format: date
        - name: endUpdateTime
          description: 结束修改时间
          in: query
          type: string
          format: date
      responses:
        '200':
          description: 返回渠道管理电话工号列表
          schema:
            $ref: '#/definitions/CrmChannelTelephoneAgentEntityList'
  /tel/crmchanneltelephoneagent/info/{id}:
    get:
      tags:
        - 渠道管理电话工号
      summary: 获取渠道管理电话工号信息
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回渠道管理电话工号信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/CrmChannelTelephoneAgentEntity'
  /tel/crmchanneltelephoneagent/save:
    post:
      tags:
        - 渠道管理电话工号
      summary: 添加渠道管理电话工号
      produces:
        - application/json
      parameters:
        - name: body
          description: 渠道管理电话工号对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CrmChannelTelephoneAgentEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /tel/crmchanneltelephoneagent/update:
    post:
      tags:
        - 渠道管理电话工号
      summary: 修改渠道管理电话工号
      produces:
        - application/json
      parameters:
        - name: body
          description: 渠道管理电话工号对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CrmChannelTelephoneAgentEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /tel/crmchanneltelephoneagent/delete:
    post:
      tags:
        - 渠道管理电话工号
      summary: 删除渠道管理电话工号
      produces:
        - application/json
      parameters:
        - name: body
          description: 渠道管理电话工号ID列表
          in: body
          type: array
          items:
            type: string
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /tel/crmchanneltelephoneagent/queryTelephoneAgentInfoByExtension:
    get:
      tags:
        - 渠道管理电话工号
      summary: 通过分机号息获取渠道管理电话工号信息
      produces:
        - application/json
      parameters:
        - name: extension
          description: 分机号
          in: query
          type: string
          required: true
        - name: agentNumber
          description: 绑定坐席工号
          in: query
          type: string
          required: true
      responses:
        '200':
          description: 返回渠道管理电话工号信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/CrmChannelTelephoneAgentVo'


  #----------------------------------------邮箱发送----------------------------------
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  PasswordForm:
    type: object
    properties:
      password:
        description: 原密码
        type: string
      newPassword:
        description: 新密码
        type: string
  SysUserEntity:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      password:
        description: 密码
        type: string
      realname:
        description: 用户真实名
        type: string
      extension:
        description: 分机号
        type: string
      isExtension:
        description: 是否绑定分机号
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      email:
        description: 邮箱
        type: string
      mobile:
        description: 手机号
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysUserEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserEntity'
  SysUserEntityEdit:
      type: object
      properties:
        userId:
          description: 用户ID
          type: integer
          format: int64
        username:
          description: 用户名
          type: string
        password:
          description: 密码
          type: string
        realname:
          description: 用户真实名
          type: string
        extension:
          description: 分机号
          type: string
        isExtension:
          description: 是否绑定分机号
          type: string
        deptId:
          description: 部门ID
          type: integer
          format: int64
        email:
          description: 邮箱
          type: string
        mobile:
          description: 手机号
          type: string
        status:
          description: 状态  0：禁用   1：正常
          type: integer
          format: int32
        roleIdList:
          description: 角色ID列表
          type: array
          items:
            type: integer
            format: int32
...
  CrmChannelTelephoneAgentEntity:
    type: object
    properties:
      id:
        description: ID
        type: string
      channelId:
        description: 渠道管理id
        type: string
      telephoneAgent:
        description: 话务工号
        type: string
      agentPassword:
        description: 工号密码
        type: string
      skillsGroup:
        description: 技能组
        type: string
      processingTime:
        description: 后续处理时长
        type: integer
        format: int32
      ringingCount:
        description: 振铃次数
        type: integer
        format: int32
      agentNumber:
        description: 绑定坐席工号
        type: string
      agentName:
        description: 绑定坐席名称
        type: string
      createCode:
        description: 创建人代码
        type: string
      createName:
        description: 创建人
        type: string
      createTime:
        description: 创建时间
        type: string
      updateCode:
        description: 修改人代码
        type: string
      updateName:
        description: 修改人
        type: string
      updateTime:
        description: 修改时间
        type: string
  CrmChannelTelephoneAgentEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/CrmChannelTelephoneAgentEntity'
  CrmChannelTelephoneAgentVo:
    type: object
    properties:
      id:
        description: ID
        type: string
      channelId:
        description: 渠道管理id
        type: string
      telephoneAgent:
        description: 话务工号
        type: string
      agentPassword:
        description: 工号密码
        type: string
      skillsGroup:
        description: 技能组
        type: string
      processingTime:
        description: 后续处理时长
        type: integer
        format: int32
      ringingCount:
        description: 振铃次数
        type: integer
        format: int32
      agentNumber:
        description: 绑定坐席工号
        type: string
      agentName:
        description: 绑定坐席名称
        type: string
      trafficBrand:
        description: 话务品牌
        type: string
      ctiAddressMaster:
        description: CTI地址主
        type: string
      ctiAddressSlave:
        description: CTI地址从
        type: string
      satisfactionNodeNumber:
        description: 满意度节点号
        type: string
      broadcasterNode:
        description: 播报工号节点
        type: string
      workingHoursBegin:
        description: 每日工作时间开始
        type: date
      workingHoursEnd:
        description: 每日工作时间结束
        type: date
      isBlacklist:
        description: 是否启用黑名单 0:否 1:是
        type: string


  #--------------------------------------邮件发送-------------------------------
...